+++ docs/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Add CISA CVSSIssuer type (OSIDB-4203)
+- Ignore modification of non-Red Hat issued CVSS scores through API
 
 ## [4.10.1] - 2025-04-30
 ### Added